improve: set default values for new properties in cached campaigns data (SDKCF-5003)
