+++ Sources/RInAppMessaging/Models/Responses/CampaignDataModels.swift
@@ -1,0 +1,71 @@
+internal struct MessagePayload: Codable {
+    let title: String
+    let messageBody: String?
+    let header: String?
+    let titleColor: String
+    let headerColor: String
+    let messageBodyColor: String
+    let backgroundColor: String
+    let frameColor: String
+    let resource: Resource
+    let messageSettings: MessageSettings
+}
+
+internal struct Resource: Codable {
+    let imageUrl: String?
+    let cropType: CampaignCropType
+}
+
+internal struct MessageSettings: Codable {
+    let displaySettings: DisplaySettings
+    let controlSettings: ControlSettings
+}
+
+internal struct DisplaySettings: Codable {
+    private enum CodingKeys: String, CodingKey {
+        case orientation
+        case slideFrom
+        case endTimeMilliseconds = "endTimeMillis"
+        case textAlign
+        case optOut
+        case html
+        case delay
+    }
+
+    let orientation: CampaignOrientation
+    let slideFrom: SlideDirection?
+    let endTimeMilliseconds: Int64
+    let textAlign: CampaignTextAlignType
+    let optOut: Bool
+    let html: Bool
+    let delay: Int
+}
+
+internal struct ControlSettings: Codable {
+    let buttons: [Button]
+    let content: Content?
+}
+
+/// For slide-up campaigns
+internal struct Content: Codable {
+    let onClickBehavior: OnClickBehavior
+    let campaignTrigger: Trigger?
+}
+
+internal struct OnClickBehavior: Codable {
+    let action: ActionType
+    let uri: String?
+}
+
+internal struct Button: Codable {
+    let buttonText: String
+    let buttonTextColor: String
+    let buttonBackgroundColor: String
+    let buttonBehavior: ButtonBehavior
+    let campaignTrigger: Trigger?
+}
+
+internal struct ButtonBehavior: Codable {
+    let action: ActionType
+    let uri: String?
+}
